refactor solution Type
